[#7616] Add function to specify the number of store in CachedStatisticsDao
