+++ collector/src/main/resources/pinpoint-collector-root.properties
@@ -124,6 +124,12 @@ cluster.listen.port=-1
 
 collector.spanEvent.sequence.limit=10000
 
+# Specifies the size to store data before flushing from CachedStatisticsDao.
+# The default is -1. If it is -1, there is no limit.
+collector.cachedStatDao.caller.limit=-1
+collector.cachedStatDao.callee.limit=-1
+collector.cachedStatDao.self.limit=-1
+
 # Flink configuration
 flink.cluster.enable=false
 flink.cluster.zookeeper.address=${pinpoint.zookeeper.address}
